Moved SignClassTransformer to the classtransformers package
